Add my initial hacks.
The `vocabulary` script can parse `*.vocab` files and print out
vocabulary questions based on the content of the file.  After the
question is printed, the program waits until something is entered on
stdin, then prints out the answer to the question.

In the future, I hope to parse the user input and execute commands such
as "do another question" or "this word was too easy/hard".  I also want
to add more vocabulary words.  I've entered about half of the words from
the "High-Frequency Word List" in my study book, but of course I want to
enter all of them.